Basic Attempts at Optimization
Thanks callgrind!

Thinking of implementing a 2nd pass redundancy check, so that
expressions such as "(if (== x x) x x)" get reduced to just "x", and so
on. Might be a fun exercise! Though would that take away from the "a
computer made this at random" aspect...?

...nah. In fact, the computer *knowing* to not do redundancies might be
even more charming!